Gui and tray updates (#94)
* Fixes to pijuice, gui and tray. Major overhaul of pijuice_cli

pijuice.py

Change duty circle -> duty cycle
Convert button event parameter from string to integer (Fixes #xx)

pijuice_gui.py
Wakeupalarm: correct day of week conversion between RTC and Raspbian
Change 'Set system time' to 'Set RTC time' since the system time is
copied
to the PiJuice RTC.

pijuice_tray.py
Correct integer divisions to Python3 convention (//). Avoids crashing
when power connected to Pi and when on Battery (Fixes #78)

pijuice_cli.py
Fixed size to 64x20 so runs nicely on standard 80x24 terminal
Use inversion for selected item throughout.
Confirmation dialog with 2 buttons use separate exits for Yes and No
Selecting No for firmware update now works.
Warning before firmware update now shows.
Suppress output for firmware update (pijuice boot output).
Waits after firmware update till firmware has restarted
Status menu now updates automatically every second
UTC time on Wakeup Alarm updates every second
Change 'Set system time' to 'Set RTC time' since the system time is
copied
to the PiJuice RTC.
Correct setting of run pin
Correct selecting input precedence (Fixes #87)
Correct selecting usb current limit and usb_micro_dpm (FIxes #87)
Button events, disable parameter for press and release
Added initial check for PiJuice and for only one instance of
pijuice_cli/pijuice_gui running.

* Fix Alarm reporting

* Do not send SIGUSR1 when pid = -1

Sending SIGUSR1 to pid -1 will send the signal to all the processes
from the user and hence terminate the users login session.

* Add python-gtk2 dependency for pijuice-gui

On recent Raspbian versions when doing an update/upgrade cycle, the
system informs that certain packages can be (auto)removed.
python-gtk2 is one of those, but needed for pijuice_tray.py.
Also updated the description field in the package control files.

+++ Software/Source/src/pijuice_gui.py
@@ -1212,7 +1212,7 @@ class PiJuiceWakeupConfig(object):
         self.currTimeLbl = Label(self.frame, textvariable=self.currTime, text="", font = "Verdana 10 bold").grid(row=1, column=0, padx=(2, 2), pady=(2, 0), columnspan = 2, sticky='W')
 
         self.setTime = StringVar()
-        self.setTimeBtn = Button(self.frame, text='Set system time', underline=0, command=lambda v=self.setTime: self._SetTime(v))
+        self.setTimeBtn = Button(self.frame, text='Set RTC time', underline=0, command=lambda v=self.setTime: self._SetTime(v))
         self.setTimeBtn.grid(row=1, column=2, padx=(10, 2), pady=(2,0), sticky = W)
 
         self.aDayOrWeekDay = IntVar()
@@ -1314,7 +1314,7 @@ class PiJuiceWakeupConfig(object):
         if t['error'] == 'NO_ERROR':
             t = t['data']
 
-            self.currTime.set(calendar.day_abbr[t['weekday']-1]+'  '+str(t['year'])+'-'+str(t['month']).rjust(2, '0')+'-'+str(t['day']).rjust(2, '0')+'  '+str(t['hour']).rjust(2, '0')+':'+str(t['minute']).rjust(2, '0')+':'+str(t['second']).rjust(2, '0')+'.'+str(t['subsecond']).rjust(2, '0'))
+            self.currTime.set(calendar.day_abbr[(t['weekday']+5) % 7]+'  '+str(t['year'])+'-'+str(t['month']).rjust(2, '0')+'-'+str(t['day']).rjust(2, '0')+'  '+str(t['hour']).rjust(2, '0')+':'+str(t['minute']).rjust(2, '0')+':'+str(t['second']).rjust(2, '0')+'.'+str(t['subsecond']).rjust(2, '0'))
         s = pijuice.rtcAlarm.GetControlStatus()
         if s['error'] == 'NO_ERROR' and s['data']['alarm_flag']:
             pijuice.rtcAlarm.ClearAlarmFlag()
@@ -1322,7 +1322,7 @@ class PiJuiceWakeupConfig(object):
 
     def _SetTime(self, v):
         t = datetime.datetime.utcnow()
-        print(pijuice.rtcAlarm.SetTime({'second':t.second, 'minute':t.minute, 'hour':t.hour, 'weekday':t.weekday()+1, 'day':t.day,  'month':t.month, 'year':t.year, 'subsecond':t.microsecond//1000000}))
+        print(pijuice.rtcAlarm.SetTime({'second':t.second, 'minute':t.minute, 'hour':t.hour, 'weekday':(t.weekday()+1) % 7 + 1, 'day':t.day,  'month':t.month, 'year':t.year, 'subsecond':t.microsecond//1000000}))
 
     def _WakeupEnableChecked(self, *args):
         ret = pijuice.rtcAlarm.SetWakeupEnabled(self.wakeupEnabled.get())
@@ -1783,10 +1783,12 @@ def PiJuiceGuiOnclosing(gui):
     gui.apply_settings()
 
     # Send signal to pijuice_tray to enable the 'Settings' menuitem again
-    try:
-        os.kill(pid, SIGUSR2)
-    except:
-        pass
+    if pid != -1:
+        # See comment on pid -1 in start_app() below
+        try:
+            os.kill(pid, SIGUSR2)
+        except:
+            pass
 
     root.destroy()
 
@@ -1818,7 +1820,7 @@ def start_app():
     # Acquire lock on lock file
     lock_file = open(LOCK_FILE, 'w')
     try:
-        fcntl.lockf(lock_file, fcntl.LOCK_EX | fcntl.LOCK_NB)
+        fcntl.flock(lock_file, fcntl.LOCK_EX | fcntl.LOCK_NB)
     except IOError:
         root.withdraw()
         MessageBox.showerror('PiJuice Settings', 'Another instance of PiJuice Settings is already running')
@@ -1830,10 +1832,13 @@ def start_app():
             pid = int(f.read())
     except:
         pid = -1
-    try:
-        os.kill(pid, SIGUSR1)
-    except:
-        pass
+    if pid != -1:
+        # Do not send SIGUSR1 to pid -1, this will kill all the processes of the user
+        # and thus terminate the login session
+        try:
+            os.kill(pid, SIGUSR1)
+        except:
+            pass
 
     root.update()
     root.minsize(400, 400)
